+++ cmd/qsctl/rm.go
@@ -5,18 +5,22 @@ import (
 	"path/filepath"
 
 	typ "github.com/Xuanwo/storage/types"
+	"github.com/c-bata/go-prompt"
 	"github.com/qingstor/noah/task"
 	"github.com/spf13/cobra"
 
+	"github.com/qingstor/qsctl/v2/cmd/qsctl/shellutils"
 	"github.com/qingstor/qsctl/v2/cmd/qsctl/taskutils"
 	"github.com/qingstor/qsctl/v2/constants"
 	"github.com/qingstor/qsctl/v2/pkg/i18n"
 	"github.com/qingstor/qsctl/v2/utils"
 )
 
-var rmInput struct {
+type rmFlags struct {
 	recursive bool
 }
+
+var rmFlag = rmFlags{}
 
 // RmCommand will handle remove object command.
 var RmCommand = &cobra.Command{
@@ -28,11 +32,18 @@ var RmCommand = &cobra.Command{
 		i18n.Sprintf("Remove objects with prefix: qsctl rm qs://bucket-name/prefix -r"),
 	),
 	Args: cobra.ExactArgs(1),
-	RunE: rmRun,
+	Run: func(cmd *cobra.Command, args []string) {
+		if err := rmRun(cmd, args); err != nil {
+			i18n.Fprintf(cmd.OutOrStderr(), "Execute %s command error: %s\n", "rm", err.Error())
+		}
+	},
+	PostRun: func(_ *cobra.Command, _ []string) {
+		rmFlag = rmFlags{}
+	},
 }
 
 func initRmFlag() {
-	RmCommand.Flags().BoolVarP(&rmInput.recursive, constants.RecursiveFlag, "r",
+	RmCommand.Flags().BoolVarP(&rmFlag.recursive, constants.RecursiveFlag, "r",
 		false, i18n.Sprintf("recursively delete keys under a specific prefix"))
 }
 
@@ -44,35 +55,26 @@ func rmRun(c *cobra.Command, args []string) (err error) {
 		return
 	}
 
-	if rootTask.GetType() == typ.ObjectTypeDir && !rmInput.recursive {
+	if rootTask.GetType() == typ.ObjectTypeDir && !rmFlag.recursive {
 		return fmt.Errorf(i18n.Sprintf("-r is required to remove a directory"))
 	}
 
-	if rmInput.recursive && rootTask.GetType() != typ.ObjectTypeDir {
+	if rmFlag.recursive && rootTask.GetType() != typ.ObjectTypeDir {
 		return fmt.Errorf(i18n.Sprintf("path should be a directory while -r is set"))
 	}
 
 	key := filepath.Join(workDir, rootTask.GetPath())
-	confirm, err := utils.CheckConfirm(i18n.Sprintf(`This operation will delete <%s>, which cannot be recovered.
-Confirm?:`, key))
-	if err != nil {
-		return
-	}
-	if !confirm {
-		return fmt.Errorf(i18n.Sprintf("Not confirmed. Object <%s> not removed.", key))
-	}
-
-	if rmInput.recursive {
+	if rmFlag.recursive {
 		t := task.NewDeleteDir(rootTask)
 		t.SetHandleObjCallback(func(o *typ.Object) {
-			fmt.Println(i18n.Sprintf("<%s> removed", o.Name))
+			i18n.Fprintf(c.OutOrStdout(), "<%s> removed\n", o.Name)
 		})
 		t.Run()
 		if t.GetFault().HasError() {
 			return t.GetFault()
 		}
 
-		i18n.Printf("Dir <%s> removed.\n", key)
+		i18n.Fprintf(c.OutOrStdout(), "Dir <%s> removed.\n", key)
 		return nil
 	}
 
@@ -82,6 +84,32 @@ Confirm?:`, key))
 		return t.GetFault()
 	}
 
-	i18n.Printf("File <%s> removed.\n", key)
+	i18n.Fprintf(c.OutOrStdout(), "File <%s> removed.\n", key)
 	return nil
+}
+
+// rmShellHandler handle rm in shell
+type rmShellHandler struct{}
+
+func (r rmShellHandler) preRunE(args []string) error {
+	err := RmCommand.Flags().Parse(args)
+	if err != nil {
+		return err
+	}
+	_, _, key, err := utils.ParseQsPath(RmCommand.Flags().Args()[0])
+	if err != nil {
+		return err
+	}
+	// try to get confirm input
+	input := prompt.Input(
+		i18n.Sprintf("confirm to remove <%s>? [y/N] ", key),
+		noSuggests)
+	if !shellutils.CheckYes(input) {
+		return fmt.Errorf(i18n.Sprintf("not confirmed"))
+	}
+	return nil
+}
+
+func (r rmShellHandler) postRun(_ error) {
+	return
 }