+++ CHANGELOG.md
@@ -4,6 +4,39 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [2.1.2] - 2020-06-30
+
+### Changed
+
+- mod: Bump storage version to adapt system specific separator to slash (#302)
+
+### Fixed
+
+- cmd/progress: Fix progress panic with negative WaitGroup counter (#296)
+
+## [2.1.1] - 2020-06-03
+
+### Changed
+
+- i18n: Set language to en-US if detect failed instead of exit 1 (#291)
+
+### Fixed
+
+- cmd/utils: Fix qsctl hang on non-interactive environment (#292)
+- Fix errors not return correctly in task exec
+
+## [2.1.0] - 2020-05-28
+
+### Added
+
+- cmd/stat: Add support to stat a bucket (#279)
+- cmd: Add global zone flag to assign zone manually (#282)
+- cmd/ls: Add support to ls buckets by long format (#280)
+
+### Fixed
+
+- cmd/progress-bar: Fix data race in progress bar (#284)
+
 ## [2.0.1] - 2020-05-21
 
 ### Changed
@@ -438,6 +471,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 - Hello, qsctl.
 
+[2.1.2]: https://github.com/qingstor/qsctl/compare/v2.1.1...v2.1.2
+[2.1.1]: https://github.com/qingstor/qsctl/compare/v2.1.0...v2.1.1
+[2.1.0]: https://github.com/qingstor/qsctl/compare/v2.0.1...v2.1.0
 [2.0.1]: https://github.com/qingstor/qsctl/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/qingstor/qsctl/compare/v2.0.0-rc.2...v2.0.0
 [2.0.0-rc.2]: https://github.com/qingstor/qsctl/compare/v2.0.0-rc.1...v2.0.0-rc.2